Remove folders and steam stuff
