+++ electron/index.js
@@ -2,7 +2,6 @@ const { app, BrowserWindow, Menu, MenuItem, ipcMain, shell, dialog, session } = 
 const path = require("path");
 const url = require("url");
 const fs = require("fs");
-const steam = require("./steam");
 const asyncLock = require("async-lock");
 const windowStateKeeper = require("electron-window-state");
 
@@ -153,7 +152,7 @@ function createWindow() {
     win.webContents.on("new-window", (event, pth) => {
         event.preventDefault();
 
-        if (pth.startsWith("https://") || pth.startsWith("steam://")) {
+        if (pth.startsWith("https://")) {
             shell.openExternal(pth);
         }
     });
@@ -378,10 +377,3 @@ try {
 ipcMain.handle("get-mods", async () => {
     return mods;
 });
-
-steam.init(isDev);
-
-// Only allow achievements and puzzle DLC if no mods are loaded
-if (mods.length === 0) {
-    steam.listen();
-}
