--- electron/steam_appid.txt
@@ -1,1 +1,0 @@
-1318690
